requirements.txt
crewai
python-dotenv
streamlit
crewai-tools
duckduckgo-search
newspaper4k
